--- src/main/java/com/hotspot/ecommerce/models/users/UserRole.java
@@ -1,8 +1,0 @@
-package com.hotspot.ecommerce.models.users;
-
-public enum UserRole {
-    USUARIO,
-    EMPRESA,
-    ADMIN
-
-}
