+++ src/main/java/com/hotspot/ecommerce/models/users/enums/UserRole.java
@@ -1,0 +1,8 @@
+package com.hotspot.ecommerce.models.users.enums;
+
+public enum UserRole {
+    Cliente,
+    Empresa,
+    Administrador
+
+}
